Cosmetics commit_hash:55b10e57134c577f52b0ad8cdca71eb54558ab66

+++ yt/yt/core/rpc/service_detail.h
@@ -1196,7 +1196,7 @@ private:
 
     // Not std::optional to guarantee lock freeness; -1 means inf.
     std::atomic<int> QueueSizeLimit_ = -1;
-    std::atomic<int> QueueByteSizeLimit_ = -1;
+    std::atomic<i64> QueueByteSizeLimit_ = -1;
     // TODO(h0pless): Add ConcurrencyLimit and ConcurrencyByteLimit.
 
     moodycamel::ConcurrentQueue<TServiceBase::TServiceContextPtr> Queue_;
